### 2018-06-19 #783 @whedon
[PRE REVIEW]: ukpopulation: UK Demographic Projections
**Submitting author:** @virgesmith (<a href="http://orcid.org/0000-0002-9951-6642">Andrew Smith</a>)
**Repository:** <a href="https://github.com/nismod/ukpopulation" target ="_blank">https://github.com/nismod/ukpopulation</a>
**Version:** 1.0.1
