### 2018-06-20 63ac949
attempt to fix git install with dep on another git package
