### 2018-06-20 d31c126
Fix test helper function names not to run as tests
