### 2018-06-20 e230aa1
updates to installation/documentation
