### 2018-06-20 3c1e914
Add essential dependencies to install_requires
