### 2018-06-20 713a49c
try fixing 3.7-dev build
