+++ .travis.yml
@@ -9,6 +9,8 @@ python:
   - "3.7-dev" # unclear if trusty supports 3.7
 
 install:
+  # workaround for 3.7-dev build error: ValueError: bad marshal data (unknown type code)
+  - pip install pycparser
   - pip install -r requirements.txt
 script:
   - ./setup.py test
