### 2018-06-20 a788d84
another try at 3.7-dev fix
