+++ .travis.yml
@@ -10,7 +10,7 @@ python:
 
 install:
   # workaround for 3.7-dev build error: ValueError: bad marshal data (unknown type code)
-  - pip install pycparser
+  - pip install distutils-pytest
   - pip install -r requirements.txt
 script:
   - ./setup.py test
