### 2018-06-20 5045071
stop allowing 3.7-dev to fail now its fixed
