### 2018-06-21 bfbd550
fix issue in python3.5 where os.path functions dont accept a Path object
