### 2018-06-29 #783 @virgesmith
Proof looks fine to me thanks
