### 2018-07-02 #783 @leeper
@soodoku Would you be interested in reviewing this for JOSS?
