### 2018-07-02 #803 @leeper
@soodoku Review checklist is here. Please open an issue on the source repo for anything that is missing, or any other changes you would like to see in the software. Let me know if you have questions. And thanks for volunteering to review!
