### 2018-07-16 d2cf720
Changes for PyPI submission #17, fix warning
