### 2018-07-16 b6b3af3
update doc - mainly installation instructions
