### 2018-07-16 787884c
update doc. Fixes #19
