### 2018-07-17 2816cfd
Fixes for conda; bump version
