+++ MANIFEST.in
@@ -1,2 +1,3 @@
+include LICENCE
 include README.md
 include doc/*
