### 2018-08-02 #803 @soodoku
@leeper: i opened a few issues around JOSS at the repo. a few weeks ago. All of them have been addressed. There are still some things I need to investigate.

For now, one q.: the version number is bumped up to 1.03 as a consequence of some of the changes i proposed. I am assuming it is ok if i check 1.01 on the checklist.
