### 2018-08-02 #803 @leeper
Yep, that's fine. Just let me know if anything else needs to be addressed. Thanks!
