### 2018-08-02 #803 @soodoku
Except for the pending 'community guidelines', everything else is all set. I offered another round of copy-edits but they are optional.
