### 2018-08-02 1c40f93
mention contributing and support #24
