### 2018-08-03 #803 @leeper
@virgesmith Looks like the package version is now 1.0.3. Can you create a new release for the current version and update your Zenodo archive? Once that's done, please let me know the updated DOI.
