### 2018-08-03 #803 @virgesmith
Updated as requested: https://zenodo.org/badge/latestdoi/131988266

DOi: 10.5281/zenodo.1328267

Thanks @soodoku @leeper!
