### 2018-08-03 #803 @arfon
@virgesmith @leeper - the compiled paper doesn't appear to have any references - is that correct? The bibtex file appears to have an entry.

You can read how to cite records in markdown [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)
